# PERF: get_block_type